requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
bandit==1.9.2
    # via -r requirements.in
black==25.11.0
    # via -r requirements.in
certifi==2025.10.5
    # via requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.0
    # via black
coverage[toml]==7.10.7
    # via pytest-cov
cryptography==46.0.2
    # via
    #   -r requirements.in
    #   pyopenssl
idna==3.11
    # via requests
iniconfig==2.1.0
    # via pytest
isort==7.0.0
    # via -r requirements.in
jsonref==1.1.0
    # via -r requirements.in
jsonschema==4.25.1
    # via -r requirements.in
jsonschema-specifications==2025.9.1
    # via jsonschema
markdown-it-py==4.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
mypy==1.18.2
    # via -r requirements.in
mypy-extensions==1.1.0
    # via
    #   black
    #   mypy
packaging==25.0
    # via
    #   black
    #   pytest
pathspec==0.12.1
    # via
    #   black
    #   mypy
    #   yamllint
platformdirs==4.5.0
    # via black
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
pycparser==2.23
    # via cffi
pygments==2.19.2
    # via
    #   pytest
    #   rich
pyopenssl==25.3.0
    # via -r requirements.in
pytest==8.4.2
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-mock
pytest-cov==7.0.0
    # via -r requirements.in
pytest-mock==3.15.1
    # via -r requirements.in
pytokens==0.3.0
    # via black
pyyaml==6.0.3
    # via
    #   -r requirements.in
    #   bandit
    #   yamllint
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via -r requirements.in
rich==14.2.0
    # via bandit
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
ruff==0.14.4
    # via -r requirements.in
types-pyyaml==6.0.12.20250915
    # via -r requirements.in
semver==3.0.4
    # via -r requirements.in
stevedore==5.6.0
    # via bandit
types-pyyaml==6.0.12.20250915
    # via -r requirements.in
types-requests==2.32.4.20250913
    # via -r requirements.in
typing-extensions==4.15.0
    # via
    #   mypy
    #   pyopenssl
    #   referencing
urllib3==2.5.0
    # via
    #   requests
    #   types-requests
yamllint==1.37.1
    # via -r requirements.in
